aplicacao/requirements.txt
asgiref==3.6.0
awsebcli==3.20.6
backports.zoneinfo==0.2.1
botocore==1.29.99
cement==2.8.2
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.3
contourpy==1.0.7
cycler==0.11.0
Django==4.1.9
django-filter==23.2
fonttools==4.39.3
idna==3.4
importlib-resources==5.12.0
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathspec==0.10.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
requests==2.26.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.2.4
termcolor==1.1.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.1.9
zipp==3.15.0
